Changes to the output Files
Just to be cautious I have added the ability for the user to split the AIJ .dat files into essentially two separate files. One of magnitudes and one of fluxes. 

This should allow for the user to use the magnitudes to find the temperature and such, but also for integration with TESS or other flux limited data. The only caveat is that the user must have allowed for the flux and magnitudes to be outputted from AIJ otherwise they will need to redo the multi-aperture photometry in AIJ for the easiest solution to this.