Create O-C_plot.py
This program takes O-C data collected from an outside source and plots them with any number of polynomial fits. This will also plot residuals with the maximum polynomial that the user enters. i.e. If the user enters polynomial value of 2 it will plot residuals of only that degree 2 and not of 1.