Calculate O-C Values
This update allows for the calculation and plotting of O-C data for TESS data only at the moment. Next update will allow for the calculation of O-C values for BSUO times of minimum. This program does NOT do any solving for ToM only calculating O-C. This also includes basic error catching for entering files and float values.