Minor Updates and Usage
This update allows the user to not have already found an Epoch value if this is their first time running this program to calculate times of minimum. The program will take the first time of minimum from the input file and use that as the Ecpoh value for the rest of the program.