Merge (#206)
* BJD to HJD Conversion (#188)

* BJD to HJD Conversion

This converts BJD to HJD more simplistically than trying to find the position of the TESS satellite, as I would assume the difference will be minimal if found.

* Removed Unused Import

* Example Master Calibration Images

Output master calibration images from the `IRAF_Reduciton.py` program.

* Hyperlink Updates (#189)

Added some hyperlinks and a sponsor section.

* Header and Overscan/Trim Fixes (#190)

The overscan subtraction and trim sections were not working correctly but they are now fixed. Also added more header key words for the DATASEC and BIASSEC, and more.

* First Attempt at Multi-Aperture Photometry (#191)

This is my first attempt at multi-aperture photometry to replace the basics of Astro ImageJ with a Python version.

* Move to the Next Day (#193)

Moves the data to the next set of data within the light curve data or to the next set of data.

* Update __init__.py

* Menu Options and Format (#194)

Changes the format of the menu to be a pandas dataframe to display the options and corresponding numbers in a table like format. This makes it easier to read than what was displayed beforehand.

* v2.8.4-alpha.7 (#195)

* Update runtime.txt

* Create .slugignore (#196)

Files and folders to ignore for Django

* Delete .slugignore

* Delete Procfile

* Update requirements.txt

* Create conf.py

* Create index.rst

* Create api.rst

* Delete settings.py

* Delete runtime.txt

* Update conf.py

* Rename README.md to README.rst

* Rename README.rst to README.md

* Delete api.rst

* Update README.md

* Create .readthedocs.yml

* Update .readthedocs.yml

* Many Errors Before Calculations (#197)

This fixes many many errors before the calculations of magnitudes even begins. Now to fix the calculations as the values are not even close to correct.

* Update .readthedocs.yml

* Update .readthedocs.yml

* Update .readthedocs.yml

* Delete jekyll-gh-pages.yml (#198)

* Rename docs/source/conf.py to docs/conf.py

* Rename docs/source/index.rst to docs/index.rst

* Update .readthedocs.yml

* Create installation.rst (#199)

* Update index.rst (#200)

* Update index.rst

* Update index.rst

* Create license.rst (#202)

* Create getting_started.rst (#201)

* Update index.rst (#203)

* Update index.rst

* Update conf.py

* Automatic Limits for Overlay (#205)

* Automatic Limits for Overlay

Found a way to get automatic limits for the overlay image for the cmap.

* Wrong Variable Name

Had the wrong variable name entered into the limit finder.

* Update __init__.py

* Update conf.py