Started Work on Using Comparisons (#212)
This starts work on creating a way to use comparison stars to calculate the magnitude of the target. I believe that I have everything up until the actual magnitude calculations. I have set the variable for "number of sky background counts per
pixel" to an arbitrary number for the time being though.