Combines TESS and AIJ into a Single Function (#313)
* Combines TESS and AIJ into a Single Function

Now that TESS magnitudes will be calculated automatically in previous processes, future researchers will need to gather the TESS magnitude data and not just the flux data. This combines what AIJ was doing with TESS data to more seamlessly meld the two together.

* Missing For Loop

Missing a for loop to loop through the number of nights the user has.