Major Improvements to Calculations (#317)
* Major Improvements to Calculations

The program now produces data that at first glance accurate to what Astro ImageJ produces. Needs to be further tested and refined before a release for others to try.

* Plot Stylings and Flux Calculations

Updates the plot styling to look a little nicer and some updates to improve the relative flux calculations.

* Comment Out Unused Variables

This comments out unused variables

* Pipeline Additions

This starts the additions of pipeline abiltiy.

* Pipeline Ability for Photemtry

Adds the pipeline ability for the continuation into multi-aperture photometry.

* Pipeline Access for Photometry

Adds in the calling of the pipeline.