Angle Fixes (#328)
* Angle Fixes

Fixes the creation of each line in the RADEC files where the angle was not actually removing stars located near the target.

* Only Use Images with Object Name

Only allow images to be used in the light curve that only contain the object name.

* Minor Updates

Minor updates to the object name and multiple_aperture inputs.

* Small Changes for Better Usage

The main fix is the plt.pause(1000) was causing the program to wait all 1000 seconds before continuing with the loop.

* Angular Distance Fixes

Changes the way comparison stars are found, and only takes ones that are within 0.2 degrees from the target star. Also fixes which comparison stars are written and not written to the RADEC files.