Memory Allocation (#332)
* Memory Allocation

This allows the user to designate a memory allocation for the IRAF reduction. Default is 16 Gb but can be changed to whatever the user wants.

* Adjusts Memory Allocation and Gain/Readout Noise

Adds memory allocation to the IRAF script and adds in gain and readout noise as things the user enters when starting the pipeline if they

* Delete IRAF_Reduction.py

* Delete pipeline.py

* Add Files to the Correct Directory

Add files to the correct directories.