Kjkoeller patch 3 (#381)
* GUI Startup

Creation of GUI for significant increase in user friendliness and ease of use. There are only 2 working scripts within this version IRAF reduction and TESS Database search.

* GUI Updates

Updates to integrate this script into the menu GUI for seamless integration

* GUI Integration

Integrating the TESS database search into the GUI

* GUI Integration

Implementing GUI integration for the TESScut script

* Allow Trim and Overscan as Inputs at GUI Level

Gives the user the opportunity to enter the Trim and Overscan regions before running the reduction process at the menu GUI

* Minor Updates and IRAF Inputs

Minor updates for easier scrollbar updating as needed and give the user the ability to enter in a trim and overscan region by plotting a bias image.

* Allow Specific Sector Selection

Allows the user to select specific sectors to download if they want or they can download all of them as usual.

* Cancel Button

Allows the user to cancel a task incase they need to change parameters or mistyped something the first time.

* Cancelling Tasks

Gives the command to cancel whatever the script is currently workling on based on if the user requests to cancel.

* Cancelling Tasks

Cancels the current task based on the user request.

* Cancelling Task

Cancels the current task based upon user request.

* APASS GUI Integration and Minor Updates

Introduces the APASS comparison star selector code into the GUI and some minor updates to the other scripts already integrated.

* Early Release of GUI

Introducing a very very early release of the GUI for some people to test.

* Incorrect Package Import

Incorrect importing of the IRAF reduction script results in immediate crashing of the package.